+++ src/OnifyTimerEventDefinition.js
@@ -1,5 +1,5 @@
-import { TimerEventDefinition } from 'bpmn-elements';
 import cronParser from 'cron-parser';
+import { TimerEventDefinition } from 'bpmn-elements';
 
 export class OnifyTimerEventDefinition extends TimerEventDefinition {
   constructor(activity, def) {
@@ -14,12 +14,14 @@ export class OnifyTimerEventDefinition extends TimerEventDefinition {
       if (cron.expressions?.length) {
         // cronParser.parseString expressions disregards seconds, so we have to parse again
         const expireAt = cronParser.parseExpression(value).next().toDate();
+
         return {
           expireAt,
           delay: expireAt - Date.now(),
         };
       }
     }
+
     return super.parse(timerType, value);
   }
 }